Nelly-Data/requirements.txt
dnspython
pymongo
requests
falcon
nltk
stanford-openie
transformers